removed all major warnings. changed most of the types to float.

+++ app/src/main/java/de/traendy/spaceshooter/engine/Entity.kt
@@ -4,7 +4,7 @@ import android.graphics.Canvas
 import android.graphics.RectF
 
 interface Entity {
-    fun updatePosition(x:Int, y:Int)
+    fun updatePosition(x:Float, y:Float)
     fun draw(canvas: Canvas)
     fun getCollisionBox():RectF
     fun isAlive():Boolean
